requirements.txt
boto3==1.10.36
botocore==1.13.36
Django==2.2.6
django-crispy-forms==1.8.0
django-storages==1.8
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
jsmin==2.2.2
Pillow==6.2.1
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
